Created service implementation for genre service
