+++ src/main/java/com/sallefy/service/GenreService.java
@@ -1,79 +1,32 @@
 package com.sallefy.service;
 
 import com.sallefy.domain.Genre;
-import com.sallefy.repository.GenreRepository;
 import com.sallefy.service.dto.GenreDTO;
-import com.sallefy.service.mapper.GenreMapper;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 
 /**
  * Service Implementation for managing {@link Genre}.
  */
-@Service
-@Transactional
-public class GenreService {
 
-    private final Logger log = LoggerFactory.getLogger(GenreService.class);
+public interface GenreService {
 
-    private final GenreRepository genreRepository;
-
-    private final GenreMapper genreMapper;
-
-    public GenreService(GenreRepository genreRepository, GenreMapper genreMapper) {
-        this.genreRepository = genreRepository;
-        this.genreMapper = genreMapper;
-    }
-
-    /**
-     * Save a genre.
-     *
-     * @param genreDTO the entity to save.
-     * @return the persisted entity.
-     */
-    public GenreDTO save(GenreDTO genreDTO) {
-        log.debug("Request to save Genre : {}", genreDTO);
-        Genre genre = genreMapper.toEntity(genreDTO);
-        genre = genreRepository.save(genre);
-        return genreMapper.toDto(genre);
-    }
+    GenreDTO save(GenreDTO genreDTO);
 
     /**
      * Get all the genres.
      *
      * @return the list of entities.
      */
-    @Transactional(readOnly = true)
-    public List<GenreDTO> findAll() {
-        log.debug("Request to get all Genres");
-        return genreRepository.findAll()
-            .stream()
-            .map(genreMapper::toDto)
-            .collect(Collectors.toCollection(LinkedList::new));
-    }
+    List<GenreDTO> findAll();
 
     /**
      * Get genres by list of Ids.
      *
      * @return the list of entities.
      */
-    @Transactional(readOnly = true)
-    public List<GenreDTO> findAllById(List<Long> genresIds) {
-        log.debug("Request to get genres by list of Ids");
-        return genreRepository.findAllById(genresIds)
-            .stream()
-            .map(genreMapper::toDto)
-            .collect(Collectors.toList());
-    }
-
+    List<GenreDTO> findAllById(List<Long> genresIds);
 
     /**
      * Get one genre by id.
@@ -81,20 +34,12 @@ public class GenreService {
      * @param id the id of the entity.
      * @return the entity.
      */
-    @Transactional(readOnly = true)
-    public Optional<GenreDTO> findOne(Long id) {
-        log.debug("Request to get Genre : {}", id);
-        return genreRepository.findById(id)
-            .map(genreMapper::toDto);
-    }
+    Optional<GenreDTO> findOne(Long id);
 
     /**
      * Delete the genre by id.
      *
      * @param id the id of the entity.
      */
-    public void delete(Long id) {
-        log.debug("Request to delete Genre : {}", id);
-        genreRepository.deleteById(id);
-    }
+    void delete(Long id);
 }
