+++ src/main/java/com/sallefy/service/exception/GenreNotFound.java
@@ -1,0 +1,8 @@
+package com.sallefy.service.exception;
+
+public class GenreNotFound extends RuntimeException {
+
+    public GenreNotFound() {
+        super("Genre not found");
+    }
+}
